+++ example/android/app/src/main/kotlin/com/pushpushgo/pushpushgo_sdk_example/MainActivity.kt
@@ -39,6 +39,7 @@ class MainActivity: FlutterActivity() {
    }
 
     override fun onNewIntent(intent: Intent) {
+       super.onNewIntent(intent)
        PushPushGoHelpers.onNewIntent(this.application, intent)
     }